+++ packages/plugin-user-memory/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @agentrail/plugin-user-memory
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [[`aac05e2`](https://github.com/yai-dev/agentrail/commit/aac05e2c4f1ce7150729010d73a4429af3f987f8)]:
+  - @agentrail/runtime-core@0.1.0
+  - @agentrail/host@0.0.4
+  - @agentrail/memo@0.0.3
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`7feaa9b`](https://github.com/yai-dev/agentrail/commit/7feaa9baca6a3e8ed2b69bf6930fd069db1a9ab7), [`f2716be`](https://github.com/yai-dev/agentrail/commit/f2716be16589f1a5d0097a258e8529cbf51fd8ba)]:
+  - @agentrail/host@0.0.3
+  - @agentrail/runtime-core@0.0.2
+  - @agentrail/memo@0.0.2
+
 ## 0.0.2
 
 ### Patch Changes